Begin moving He3 bits to their own file
