+++ src/ess/polarization/__init__.py
@@ -11,36 +11,18 @@ except importlib.metadata.PackageNotFoundError:
 
 del importlib
 
+
 from .base import (
-    Analyzer,
-    Cell,
     CellInBeamLog,
     CellSpinLog,
-    Depolarized,
-    DirectBeamBackgroundQRange,
-    DirectBeamNoCell,
-    DirectBeamQRange,
     Down,
-    He3CellLength,
-    He3CellPressure,
-    He3DirectBeam,
-    He3FillingTime,
-    He3Opacity0,
-    He3OpacityFunction,
-    He3PolarizationFunction,
-    He3TransmissionEmptyGlass,
-    He3TransmissionFunction,
     PolarizationCorrectedSampleData,
-    Polarized,
-    Polarizer,
     RunSectionLog,
     SampleInBeamLog,
     Up,
     WavelengthBins,
     correct_sample_data_for_polarization,
     determine_run_section,
-    direct_beam,
-    direct_beam_with_cell,
     extract_analyzer_direct_beam_polarized,
     extract_direct_beam,
     extract_polarizer_direct_beam_polarized,
@@ -48,55 +30,76 @@ from .base import (
     extract_sample_data_down_up,
     extract_sample_data_up_down,
     extract_sample_data_up_up,
+    providers,
+    run_reduction_workflow,
+)
+from .he3 import (
+    Analyzer,
+    Cell,
+    Depolarized,
+    DirectBeamBackgroundQRange,
+    DirectBeamNoCell,
+    DirectBeamQRange,
+    He3CellLength,
+    He3CellPressure,
+    He3FillingTime,
+    He3Opacity0,
+    He3OpacityFunction,
+    He3PolarizationFunction,
+    He3TransmissionEmptyGlass,
+    He3TransmissionFunction,
+    Polarized,
+    Polarizer,
+    direct_beam,
+    direct_beam_with_cell,
     get_he3_transmission_from_fit_to_direct_beam,
     he3_opacity_from_cell_params,
     he3_opacity_function_from_beam_data,
     he3_opacity_function_from_cell_opacity,
-    providers,
-    run_reduction_workflow,
 )
 
 __all__ = [
-    "providers",
+    "Analyzer",
+    "Cell",
+    "CellInBeamLog",
+    "CellSpinLog",
+    "correct_sample_data_for_polarization",
+    "Depolarized",
     "detect_run_section",
-    "run_reduction_workflow",
     "determine_run_section",
     "direct_beam",
+    "DirectBeamBackgroundQRange",
+    "DirectBeamNoCell",
+    "DirectBeamQRange",
     "direct_beam_with_cell",
+    "Down",
+    "extract_analyzer_direct_beam_polarized",
     "extract_direct_beam",
     "extract_polarizer_direct_beam_polarized",
-    "extract_analyzer_direct_beam_polarized",
     "extract_sample_data_down_down",
     "extract_sample_data_down_up",
     "extract_sample_data_up_down",
     "extract_sample_data_up_up",
+    "get_he3_transmission_from_fit_to_direct_beam",
+    "He3CellLength",
+    "He3CellPressure",
+    "He3DirectBeam",
+    "He3FillingTime",
+    "He3Opacity0",
+    "he3_opacity_from_cell_params",
+    "He3OpacityFunction",
     "he3_opacity_function_from_beam_data",
     "he3_opacity_function_from_cell_opacity",
-    "he3_opacity_from_cell_params",
-    "get_he3_transmission_from_fit_to_direct_beam",
-    "correct_sample_data_for_polarization",
-    "Up",
-    "Down",
-    "Depolarized",
-    "Polarized",
-    "Analyzer",
-    "Polarizer",
-    "Cell",
-    "WavelengthBins",
-    "DirectBeamQRange",
-    "DirectBeamBackgroundQRange",
     "He3PolarizationFunction",
-    "He3TransmissionFunction",
-    "He3CellPressure",
-    "He3CellLength",
-    "He3FillingTime",
-    "He3OpacityFunction",
     "He3TransmissionEmptyGlass",
-    "DirectBeamNoCell",
-    "He3DirectBeam",
+    "He3TransmissionFunction",
     "PolarizationCorrectedSampleData",
-    "CellSpinLog",
+    "Polarized",
+    "Polarizer",
+    "providers",
+    "run_reduction_workflow",
     "RunSectionLog",
     "SampleInBeamLog",
-    "CellInBeamLog",
+    "Up",
+    "WavelengthBins",
 ]
